@@ -1,15 +1,25 @@
 {
   "name": "react-avatar",
-  "version": "2.4.3",
+  "version": "3.4.6",
   "description": "Universal React avatar component makes it possible to generate avatars based on user information.",
   "main": "lib/index.js",
+  "module": "es/index.js",
+  "types": "index.d.ts",
+  "typings": "index.d.ts",
   "scripts": {
-    "clean": "rm -rf ./lib/*",
-    "transpile": "babel ./src --out-dir ./lib",
-    "demo": "browserify ./lib/demo.js -o ./build/demo.js --transform browserify-global-shim --standalone ReactAvatarDemo",
-    "dev": "npm run clean && npm run transpile && npm run demo",
-    "test": "eslint src/index.js",
-    "serve": "python -m SimpleHTTPServer"
+    "prepack": "npm -s run test && npm -s run build",
+    "clean": "rm -rf ./{lib,es}/*",
+    "build": "npm -s run clean && npm -s run build:commonjs && npm -s run build:modules",
+    "build:commonjs": "BABEL_MODULES=commonjs babel ./src --out-dir ./lib",
+    "build:modules": "babel ./src --out-dir ./es",
+    "build:demo": "webpack -p",
+    "demo": "npm -s run build:demo",
+    "dev": "webpack-dev-server",
+    "serve": "npm -s run dev",
+    "test": "npm -s run test:lint",
+    "test:lint": "eslint src",
+    "postpublish": "npm -s run publish:docs",
+    "publish:docs": "npm -s run build:demo && gh-pages -d build"
   },
   "license": "MIT",
   "repository": {
@@ -26,47 +36,36 @@
     "react-component",
     "avatar"
   ],
-  "homepage": "https://github.com/sitebase/react-avatar",
+  "homepage": "https://www.sitebase.be/react-avatar/",
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0 || ^16.0.0",
-    "react-addons-shallow-compare": "^0.14.0 || ^15.0.0 || ^16.0.0",
-    "prop-types": "^0.14.0 || ^15.0.0 || ^16.0.0"
+    "react": "^15.0.0 || ^16.0.0",
+    "prop-types": "^15.0.0 || ^16.0.0"
   },
-  "browserify-global-shim": {},
   "devDependencies": {
-    "babel": "6.5.0",
-    "babel-cli": "^6.24.1",
-    "babel-core": "6.5.0",
+    "@babel/cli": "^7.0.0",
+    "@babel/core": "^7.0.1",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@types/react": "^15.0.0",
     "babel-eslint": "^6.0.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-dev": "^1.0.0",
-    "babel-plugin-dev-expression": "^0.1.0",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-inline-environment-variables": "^6.5.0",
-    "babel-plugin-transform-member-expression-literals": "^6.5.0",
-    "babel-plugin-transform-property-literals": "^6.5.0",
-    "babel-plugin-transform-react-constant-elements": "^6.5.0",
-    "babel-plugin-transform-react-inline-elements": "^6.5.0",
-    "babel-plugin-transform-runtime": "^6.5.2",
-    "babel-plugin-transform-undefined-to-void": "^6.5.0",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-regenerator-runtime": "^6.5.0",
-    "babel-register": "^6.5.2",
-    "babelify": "^7.2.0",
-    "browserify": "^13.0.1",
-    "browserify-global-shim": "^1.0.3",
-    "eslint": "1.10.3",
-    "eslint-plugin-react": "3.5.1",
+    "babel-loader": "^8.0.2",
+    "eslint": "^1.10.3",
+    "eslint-loader": "^2.1.0",
+    "eslint-plugin-react": "^4.3.0",
+    "file-loader": "^2.0.0",
+    "gh-pages": "^1.2.0",
     "react": "^15.0.0",
     "react-dom": "^15.0.0",
-    "react-addons-shallow-compare": "^15.0.0",
-    "reactify": "^1.0.0",
-    "watchify": "^3.7.0"
+    "webpack": "^4.19.0",
+    "webpack-bundle-analyzer": "^3.0.2",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.8"
   },
   "dependencies": {
     "babel-runtime": ">=5.0.0",
+    "core-js": "^2.5.7",
+    "is-retina": "^1.0.3",
     "md5": "^2.0.0"
   }
 }
