@@ -1,24 +1,29 @@
 'use strict';
 
-import React from 'react';
+import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
-import shallowCompare from 'react-addons-shallow-compare';
-import {cacheFailingSource, hasSourceFailedBefore} from './utils.js';
 
-import gravatarSource from './sources/Gravatar.js';
-import facebookSource from './sources/Facebook.js';
-import vkontakteSource from './sources/Vkontakte.js';
-import twitterSource from './sources/Twitter.js';
-import googleSource from './sources/Google.js';
-import skypeSource from './sources/Skype.js';
-import valueSource from './sources/Value.js';
-import srcSource from './sources/Src.js';
-import iconSource from './sources/Icon.js';
+import {withConfig, ConfigProvider} from './context';
+import {getRandomColor, parseSize} from './utils';
+import InternalState from './internal-state';
+
+import gravatarSource from './sources/Gravatar';
+import facebookSource from './sources/Facebook';
+import vkontakteSource from './sources/Vkontakte';
+import twitterSource from './sources/Twitter';
+import googleSource from './sources/Google';
+import skypeSource from './sources/Skype';
+import valueSource from './sources/Value';
+import srcSource from './sources/Src';
+import iconSource from './sources/Icon';
+import redirectSource from './sources/AvatarRedirect';
 
 const SOURCES = [
     facebookSource,
     googleSource,
     twitterSource,
+    redirectSource('twitter', 'twitterHandle'),
+    redirectSource('instagram', 'instagramId'),
     vkontakteSource,
     skypeSource,
     gravatarSource,
@@ -27,14 +32,44 @@ const SOURCES = [
     iconSource
 ];
 
-export default class Avatar extends React.Component {
+// Collect propTypes for each individual source
+const sourcePropTypes = SOURCES.reduce((r, s) => Object.assign(r, s.propTypes), {});
+
+export {getRandomColor} from './utils';
+export {ConfigProvider} from './context';
+
+function matchSource(Source, props, cb) {
+    const { cache } = props;
+    const instance = new Source(props);
+
+    if(!instance.isCompatible(props))
+        return cb();
+
+    instance.get((state) => {
+        const failedBefore = state &&
+            state.hasOwnProperty('src') &&
+            cache.hasSourceFailedBefore(state.src);
+
+        if(!failedBefore && state) {
+            cb(state);
+        } else {
+            cb();
+        }
+    });
+}
+
+export
+class Avatar extends PureComponent {
+
     static displayName = 'Avatar'
+
     static propTypes = {
+        // PropTypes defined on sources
+        ...sourcePropTypes,
+
         className: PropTypes.string,
-        innerClassName: PropTypes.string,
         fgColor: PropTypes.string,
         color: PropTypes.string,
-        colors: PropTypes.array,
         name: PropTypes.string,
         value: PropTypes.string,
         email: PropTypes.string,
@@ -46,217 +81,249 @@ export default class Avatar extends React.Component {
         twitterHandle: PropTypes.string,
         vkontakteId: PropTypes.string,
         skypeId: PropTypes.string,
-        round: PropTypes.bool,
+        colors: PropTypes.arrayOf(PropTypes.string),
+        round: PropTypes.oneOfType([
+            PropTypes.bool,
+            PropTypes.string
+        ]),
         style: PropTypes.object,
-        size: PropTypes.number,
+        size: PropTypes.oneOfType([
+            PropTypes.number,
+            PropTypes.string
+        ]),
         textSizeRatio: PropTypes.number,
-        unstyled: PropTypes.bool
+        unstyled: PropTypes.bool,
+        cache: PropTypes.object,
+        onClick: PropTypes.func
+
     }
 
     static defaultProps = {
         className: 'sb-avatar',
         innerClassName:'',
         fgColor: '#FFF',
-        color: null,
-        name: null,
-        value: null,
-        email: null,
-        md5Email: null,
-        facebookId: null,
-        googleId: null,
-        twitterHandle: null,
-        vkontakteId: null,
-        skypeId: null,
         round: false,
         size: 100,
-        style: null,
         textSizeRatio: 3,
         unstyled: false
     }
 
     constructor(props) {
         super(props);
+
         this.state = {
-            _internal: {
-                sourcePointer: 0
-            },
-            src: props.src,
-            srcSet: props.srcSet,
+            src: null,
+            srcSet: null,
+            internal: null,
             value: null,
             color: props.color
         };
     }
 
-    componentWillMount() {
+    componentDidMount() {
         this.fetch();
     }
 
     componentWillReceiveProps(newProps) {
-        const nextState = {};
-        if (newProps.src !== this.props.src)
-            nextState.src = newProps.src;
+        let needsUpdate = false;
 
-        if (newProps.srcSet !== this.props.srcSet)
-            nextState.srcSet = newProps.srcSet;
+        // This seems redundant
+        //
+        // Props that need to be in `state` are
+        // `value`, `src` and `color`
+        for (const prop in sourcePropTypes)
+            needsUpdate = needsUpdate || (newProps[prop] !== this.props[prop]);
 
-        if (newProps.name !== this.props.name)
-            nextState.name = newProps.name;
+        // if ((newProps.srcSet !== this.props.srcSet) || (newProps.name !== this.props.name)) {
+        //     needsUpdate = true
+        // }
 
-        if (newProps.value !== this.props.value)
-            nextState.value = newProps.value;
+        if (needsUpdate)
+            setTimeout(this.fetch, 0);
+    }
 
-        if (newProps.email !== this.props.email)
-            nextState.email = newProps.email;
-
-        if (newProps.md5Email !== this.props.md5Email)
-            nextState.md5Email = newProps.md5Email;
-
-        if (newProps.facebookId !== this.props.facebookId)
-            nextState.facebookId = newProps.facebookId;
-
-        if (newProps.googleId !== this.props.googleId)
-            nextState.googleId = newProps.googleId;
-
-        if (newProps.twitterHandle !== this.props.twitterHandle)
-            nextState.twitterHandle = newProps.twitterHandle;
-
-        if (newProps.vkontakteId !== this.props.vkontakteId)
-            nextState.vkontakteId = newProps.vkontakteId;
-
-        if (newProps.skypeId !== this.props.skypeId)
-            nextState.skypeId = newProps.skypeId;
-
-
-        if(Object.keys(nextState) !== 0) {
-            nextState._internal = this.state._internal;
-            nextState._internal.sourcePointer = 0;
-            this.setState(nextState, this.fetch);
+    componentWillUnmount() {
+        if (this.state.internal) {
+            this.state.internal.active = false;
         }
     }
 
-    shouldComponentUpdate(nextProps, nextState) {
-        return shallowCompare(this, nextProps, nextState);
-    }
+    static getRandomColor = getRandomColor
+    static ConfigProvider = ConfigProvider
 
-    tryNextsource = (Source, next) => {
+    _createFetcher = (internal) => (errEvent) => {
+        const { cache } = this.props;
 
-        const instance = new Source(this.props);
+        if (!internal.isActive(this.state))
+            return;
 
-        if(!instance.isCompatible(this.props))
-            return next();
+        // Mark img source as failed for future reference
+        if( errEvent && errEvent.type === 'error' )
+            cache.sourceFailed(errEvent.target.src);
 
-        instance.get((state) => {
-            const failedBefore = state &&
-                state.hasOwnProperty('src') &&
-                hasSourceFailedBefore(state.src);
+        const pointer = internal.sourcePointer;
+        if(SOURCES.length === pointer)
+            return;
 
-            if(!failedBefore && state) {
-                // console.log(state);
-                this.setState(state);
-            } else {
-                next();
-            }
-        });
-    };
+        const source = SOURCES[pointer];
 
-    fetch = (event) => {
-        // If fetch was triggered by img onError
-        // then set state src back to null so render will
-        // automatically switch a text avatar if there is no
-        // other social ID available to try
-        if( event && event.type === 'error' ) {
-            cacheFailingSource(this.state.src);
-            this.setState({src: null});
-        }
+        internal.sourcePointer++;
 
-        // console.log('## fetch');
+        matchSource(source, this.props, (nextState) => {
+            if (!nextState)
+                return setTimeout(internal.fetch, 0);
 
-        const id = this._fetchId = this._fetchId ? this._fetchId + 1 : 1;
-
-        var tryFetch = () => {
-            if(SOURCES.length === this.state._internal.sourcePointer)
+            if (!internal.isActive(this.state))
                 return;
 
-            const source = SOURCES[this.state._internal.sourcePointer];
+            // Reset other values to prevent them from sticking (#51)
+            nextState = {
+                src: null,
+                value: null,
+                color: null,
+                name: null,
+                srcSet: null,
 
-            const internal = this.state._internal;
-            internal.sourcePointer++;
+                ...nextState
+            };
 
-            // console.log('## try fetch', id, this._fetchId, internal.sourcePointer-1);
-            this.setState({
-                _internal: internal
-            }, () => {
-                this.tryNextsource(source, () => {
-                    // console.log('-- next', id, this._fetchId);
-                    if (id === this._fetchId) {
-                        tryFetch();
-                    }
-                });
+            this.setState(state => {
+                // Internal state has been reset => we received new props
+                return internal.isActive(state) ? nextState : {};
             });
-        };
+        });
+    }
 
-        tryFetch();
+    fetch = () => {
+        const internal = new InternalState();
+        internal.fetch = this._createFetcher(internal);
 
+        this.setState({ internal }, internal.fetch);
     };
 
+    _scaleTextNode = (node) => {
+        const { unstyled, textSizeRatio } = this.props;
+
+        if (!node || unstyled) return;
+
+
+        const parent = node.parentNode;
+
+        // Reset font-size such that scaling works correctly (#133)
+        parent.style.fontSize = null;
+
+        const textWidth = node.getBoundingClientRect().width;
+        if (textWidth < 0)
+            return;
+
+        const containerWidth = parent.getBoundingClientRect().width;
+        const ratio = containerWidth / textWidth;
+
+        // Set font-size on parent span, otherwise the `table-cell` span
+        // will cause alignment issues.
+        parent.style.fontSize = `calc((100% * ${ratio}) / ${textSizeRatio})`;
+    }
+
     _renderAsImage() {
-        const size = this.props.size;
-        const round = this.props.round;
-        const alt = this.props.name || this.props.value;
-        const imageStyle = this.props.unstyled ? null : {
+        const { className, round, unstyled, name, value } = this.props;
+        const { internal } = this.state;
+        const size = parseSize(this.props.size);
+        const alt = name || value;
+
+        const imageStyle = unstyled ? null : {
             maxWidth: '100%',
-            width: size,
-            height: size,
-            borderRadius: (round ? 500 : 0)
+            width: size.str,
+            height: size.str,
+            borderRadius: (round === true ? '100%' : round)
         };
+
         return (
-            <img className={this.props.innerClassName}
-                width={this.props.size}
-                height={this.props.size}
+            <img className={className + ' sb-avatar__image'}
+                width={size.str}
+                height={size.str}
                 style={imageStyle}
                 src={this.state.src}
                 srcSet={this.state.srcSet}
                 alt={alt}
-                onError={this.fetch} />
+                onError={internal && internal.fetch} />
         );
     }
 
     _renderAsText() {
-        const size = this.props.size;
-        const textSizeRatio = this.props.textSizeRatio;
-        const round = this.props.round;
-        const initialsStyle = this.props.unstyled ? null : {
-            width: size,
-            height: size,
-            font: Math.floor(size / textSizeRatio) + 'px Helvetica, Arial, sans-serif',
-            lineHeight: size + 'px', // yes, px suffix is needed on lineHeight
+        const { className, round, unstyled } = this.props;
+        const size = parseSize(this.props.size);
+
+        const initialsStyle = unstyled ? null : {
+            width: size.str,
+            height: size.str,
+            lineHeight: 'initial',
             textAlign: 'center',
             textTransform: 'uppercase',
             color: this.props.fgColor,
             background: this.state.color,
-            borderRadius: (round ? '100%' : 0)
+            borderRadius: (round === true ? '100%' : round)
         };
+
+        const tableStyle = unstyled ? null : {
+            display: 'table',
+            width: '100%',
+            height: '100%'
+        };
+
+        const spanStyle = unstyled ? null : {
+            display: 'table-cell',
+            verticalAlign: 'middle'
+        };
+
         return (
-            <div style={initialsStyle}>
-                {this.state.value}
+            <div className={className + ' sb-avatar__text'}
+                style={initialsStyle}>
+                <div style={tableStyle}>
+                    <span style={spanStyle}>
+                        <span ref={this._scaleTextNode} key={this.state.value}>
+                            {this.state.value}
+                        </span>
+                    </span>
+                </div>
             </div>
         );
     }
 
     render() {
-        const size = this.props.size;
-        const hostStyle = this.props.unstyled ? null : {
+        const { className, unstyled, round, style, onClick } = this.props;
+        const { src, sourceName } = this.state;
+        const size = parseSize(this.props.size);
+
+        const hostStyle = unstyled ? null : {
             display: 'inline-block',
-            width: size,
-            height: size,
-            borderRadius: (this.props.round ? 500 : 0),
-            ...this.props.style
+            verticalAlign: 'middle',
+            width: size.str,
+            height: size.str,
+            borderRadius: (round === true ? '100%' : round),
+            fontFamily: 'Helvetica, Arial, sans-serif',
+            ...style
         };
+
+        const classNames = [ className, 'sb-avatar' ];
+
+        if (sourceName) {
+            const source = sourceName.toLowerCase()
+                .replace(/[^a-z0-9-]+/g, '-') // only allow alphanumeric
+                .replace(/^-+|-+$/g, ''); // trim `-`
+            classNames.push('sb-avatar--' + source);
+        }
+
         return (
-            <div className={this.props.className}
+            <div className={classNames.join(' ')}
+                onClick={onClick}
                 style={hostStyle}>
-                {this.state.src ? this._renderAsImage() : this._renderAsText()}
+                {src ? this._renderAsImage() : this._renderAsText()}
             </div>
         );
     }
 }
+
+export default Object.assign(withConfig(Avatar), {
+    getRandomColor,
+    ConfigProvider
+});
