@@ -1,317 +1,379 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-    value: true
+  value: true
 });
+Object.defineProperty(exports, "ConfigProvider", {
+  enumerable: true,
+  get: function get() {
+    return _context.ConfigProvider;
+  }
+});
+Object.defineProperty(exports, "getRandomColor", {
+  enumerable: true,
+  get: function get() {
+    return _utils.getRandomColor;
+  }
+});
+exports.default = exports.Avatar = void 0;
 
-var _extends2 = require('babel-runtime/helpers/extends');
+require("core-js/modules/es7.symbol.async-iterator");
 
-var _extends3 = _interopRequireDefault(_extends2);
+require("core-js/modules/es6.symbol");
 
-var _keys = require('babel-runtime/core-js/object/keys');
+require("core-js/modules/web.dom.iterable");
 
-var _keys2 = _interopRequireDefault(_keys);
+require("core-js/modules/es6.array.iterator");
 
-var _getPrototypeOf = require('babel-runtime/core-js/object/get-prototype-of');
+require("core-js/modules/es6.object.keys");
 
-var _getPrototypeOf2 = _interopRequireDefault(_getPrototypeOf);
+require("core-js/modules/es6.object.set-prototype-of");
 
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+require("core-js/modules/es6.regexp.replace");
 
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+require("core-js/modules/es6.function.name");
 
-var _createClass2 = require('babel-runtime/helpers/createClass');
+require("core-js/modules/es6.object.assign");
 
-var _createClass3 = _interopRequireDefault(_createClass2);
+var _react = _interopRequireWildcard(require("react"));
 
-var _possibleConstructorReturn2 = require('babel-runtime/helpers/possibleConstructorReturn');
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _possibleConstructorReturn3 = _interopRequireDefault(_possibleConstructorReturn2);
+var _context = require("./context");
 
-var _inherits2 = require('babel-runtime/helpers/inherits');
+var _utils = require("./utils");
 
-var _inherits3 = _interopRequireDefault(_inherits2);
+var _internalState = _interopRequireDefault(require("./internal-state"));
 
-var _react = require('react');
+var _Gravatar = _interopRequireDefault(require("./sources/Gravatar"));
 
-var _react2 = _interopRequireDefault(_react);
+var _Facebook = _interopRequireDefault(require("./sources/Facebook"));
 
-var _propTypes = require('prop-types');
+var _Vkontakte = _interopRequireDefault(require("./sources/Vkontakte"));
 
-var _propTypes2 = _interopRequireDefault(_propTypes);
+var _Twitter = _interopRequireDefault(require("./sources/Twitter"));
 
-var _reactAddonsShallowCompare = require('react-addons-shallow-compare');
+var _Google = _interopRequireDefault(require("./sources/Google"));
 
-var _reactAddonsShallowCompare2 = _interopRequireDefault(_reactAddonsShallowCompare);
+var _Skype = _interopRequireDefault(require("./sources/Skype"));
 
-var _utils = require('./utils.js');
+var _Value = _interopRequireDefault(require("./sources/Value"));
 
-var _Gravatar = require('./sources/Gravatar.js');
+var _Src = _interopRequireDefault(require("./sources/Src"));
 
-var _Gravatar2 = _interopRequireDefault(_Gravatar);
+var _Icon = _interopRequireDefault(require("./sources/Icon"));
 
-var _Facebook = require('./sources/Facebook.js');
-
-var _Facebook2 = _interopRequireDefault(_Facebook);
-
-var _Vkontakte = require('./sources/Vkontakte.js');
-
-var _Vkontakte2 = _interopRequireDefault(_Vkontakte);
-
-var _Twitter = require('./sources/Twitter.js');
-
-var _Twitter2 = _interopRequireDefault(_Twitter);
-
-var _Google = require('./sources/Google.js');
-
-var _Google2 = _interopRequireDefault(_Google);
-
-var _Skype = require('./sources/Skype.js');
-
-var _Skype2 = _interopRequireDefault(_Skype);
-
-var _Value = require('./sources/Value.js');
-
-var _Value2 = _interopRequireDefault(_Value);
-
-var _Src = require('./sources/Src.js');
-
-var _Src2 = _interopRequireDefault(_Src);
-
-var _Icon = require('./sources/Icon.js');
-
-var _Icon2 = _interopRequireDefault(_Icon);
+var _AvatarRedirect = _interopRequireDefault(require("./sources/AvatarRedirect"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var SOURCES = [_Facebook2.default, _Google2.default, _Twitter2.default, _Vkontakte2.default, _Skype2.default, _Gravatar2.default, _Src2.default, _Value2.default, _Icon2.default];
+function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj.default = obj; return newObj; } }
 
-var Avatar = function (_React$Component) {
-    (0, _inherits3.default)(Avatar, _React$Component);
+function _typeof(obj) { if (typeof Symbol === "function" && typeof Symbol.iterator === "symbol") { _typeof = function _typeof(obj) { return typeof obj; }; } else { _typeof = function _typeof(obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; }; } return _typeof(obj); }
 
-    function Avatar(props) {
-        (0, _classCallCheck3.default)(this, Avatar);
+function _objectSpread(target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i] != null ? arguments[i] : {}; var ownKeys = Object.keys(source); if (typeof Object.getOwnPropertySymbols === 'function') { ownKeys = ownKeys.concat(Object.getOwnPropertySymbols(source).filter(function (sym) { return Object.getOwnPropertyDescriptor(source, sym).enumerable; })); } ownKeys.forEach(function (key) { _defineProperty(target, key, source[key]); }); } return target; }
 
-        var _this = (0, _possibleConstructorReturn3.default)(this, (Avatar.__proto__ || (0, _getPrototypeOf2.default)(Avatar)).call(this, props));
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-        _this.tryNextsource = function (Source, next) {
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
 
-            var instance = new Source(_this.props);
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
 
-            if (!instance.isCompatible(_this.props)) return next();
+function _possibleConstructorReturn(self, call) { if (call && (_typeof(call) === "object" || typeof call === "function")) { return call; } return _assertThisInitialized(self); }
 
-            instance.get(function (state) {
-                var failedBefore = state && state.hasOwnProperty('src') && (0, _utils.hasSourceFailedBefore)(state.src);
+function _getPrototypeOf(o) { _getPrototypeOf = Object.setPrototypeOf ? Object.getPrototypeOf : function _getPrototypeOf(o) { return o.__proto__ || Object.getPrototypeOf(o); }; return _getPrototypeOf(o); }
 
-                if (!failedBefore && state) {
-                    // console.log(state);
-                    _this.setState(state);
-                } else {
-                    next();
-                }
-            });
-        };
+function _inherits(subClass, superClass) { if (typeof superClass !== "function" && superClass !== null) { throw new TypeError("Super expression must either be null or a function"); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, writable: true, configurable: true } }); if (superClass) _setPrototypeOf(subClass, superClass); }
 
-        _this.fetch = function (event) {
-            // If fetch was triggered by img onError
-            // then set state src back to null so render will
-            // automatically switch a text avatar if there is no
-            // other social ID available to try
-            if (event && event.type === 'error') {
-                (0, _utils.cacheFailingSource)(_this.state.src);
-                _this.setState({ src: null });
-            }
+function _setPrototypeOf(o, p) { _setPrototypeOf = Object.setPrototypeOf || function _setPrototypeOf(o, p) { o.__proto__ = p; return o; }; return _setPrototypeOf(o, p); }
 
-            // console.log('## fetch');
+function _assertThisInitialized(self) { if (self === void 0) { throw new ReferenceError("this hasn't been initialised - super() hasn't been called"); } return self; }
 
-            var id = _this._fetchId = _this._fetchId ? _this._fetchId + 1 : 1;
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
-            var tryFetch = function tryFetch() {
-                if (SOURCES.length === _this.state._internal.sourcePointer) return;
+var SOURCES = [_Facebook.default, _Google.default, _Twitter.default, (0, _AvatarRedirect.default)('twitter', 'twitterHandle'), (0, _AvatarRedirect.default)('instagram', 'instagramId'), _Vkontakte.default, _Skype.default, _Gravatar.default, _Src.default, _Value.default, _Icon.default]; // Collect propTypes for each individual source
 
-                var source = SOURCES[_this.state._internal.sourcePointer];
+var sourcePropTypes = SOURCES.reduce(function (r, s) {
+  return Object.assign(r, s.propTypes);
+}, {});
 
-                var internal = _this.state._internal;
-                internal.sourcePointer++;
+function matchSource(Source, props, cb) {
+  var cache = props.cache;
+  var instance = new Source(props);
+  if (!instance.isCompatible(props)) return cb();
+  instance.get(function (state) {
+    var failedBefore = state && state.hasOwnProperty('src') && cache.hasSourceFailedBefore(state.src);
 
-                // console.log('## try fetch', id, this._fetchId, internal.sourcePointer-1);
-                _this.setState({
-                    _internal: internal
-                }, function () {
-                    _this.tryNextsource(source, function () {
-                        // console.log('-- next', id, this._fetchId);
-                        if (id === _this._fetchId) {
-                            tryFetch();
-                        }
-                    });
-                });
-            };
-
-            tryFetch();
-        };
-
-        _this.state = {
-            _internal: {
-                sourcePointer: 0
-            },
-            src: props.src,
-            srcSet: props.srcSet,
-            value: null,
-            color: props.color
-        };
-        return _this;
+    if (!failedBefore && state) {
+      cb(state);
+    } else {
+      cb();
     }
+  });
+}
 
-    (0, _createClass3.default)(Avatar, [{
-        key: 'componentWillMount',
-        value: function componentWillMount() {
-            this.fetch();
-        }
-    }, {
-        key: 'componentWillReceiveProps',
-        value: function componentWillReceiveProps(newProps) {
-            var nextState = {};
-            if (newProps.src !== this.props.src) nextState.src = newProps.src;
+var Avatar =
+/*#__PURE__*/
+function (_PureComponent) {
+  _inherits(Avatar, _PureComponent);
 
-            if (newProps.srcSet !== this.props.srcSet) nextState.srcSet = newProps.srcSet;
+  function Avatar(props) {
+    var _this;
 
-            if (newProps.name !== this.props.name) nextState.name = newProps.name;
+    _classCallCheck(this, Avatar);
 
-            if (newProps.value !== this.props.value) nextState.value = newProps.value;
+    _this = _possibleConstructorReturn(this, _getPrototypeOf(Avatar).call(this, props));
 
-            if (newProps.email !== this.props.email) nextState.email = newProps.email;
+    _defineProperty(_assertThisInitialized(_assertThisInitialized(_this)), "_createFetcher", function (internal) {
+      return function (errEvent) {
+        var cache = _this.props.cache;
+        if (!internal.isActive(_this.state)) return; // Mark img source as failed for future reference
 
-            if (newProps.md5Email !== this.props.md5Email) nextState.md5Email = newProps.md5Email;
+        if (errEvent && errEvent.type === 'error') cache.sourceFailed(errEvent.target.src);
+        var pointer = internal.sourcePointer;
+        if (SOURCES.length === pointer) return;
+        var source = SOURCES[pointer];
+        internal.sourcePointer++;
+        matchSource(source, _this.props, function (nextState) {
+          if (!nextState) return setTimeout(internal.fetch, 0);
+          if (!internal.isActive(_this.state)) return; // Reset other values to prevent them from sticking (#51)
 
-            if (newProps.facebookId !== this.props.facebookId) nextState.facebookId = newProps.facebookId;
+          nextState = _objectSpread({
+            src: null,
+            value: null,
+            color: null,
+            name: null,
+            srcSet: null
+          }, nextState);
 
-            if (newProps.googleId !== this.props.googleId) nextState.googleId = newProps.googleId;
+          _this.setState(function (state) {
+            // Internal state has been reset => we received new props
+            return internal.isActive(state) ? nextState : {};
+          });
+        });
+      };
+    });
 
-            if (newProps.twitterHandle !== this.props.twitterHandle) nextState.twitterHandle = newProps.twitterHandle;
+    _defineProperty(_assertThisInitialized(_assertThisInitialized(_this)), "fetch", function () {
+      var internal = new _internalState.default();
+      internal.fetch = _this._createFetcher(internal);
 
-            if (newProps.vkontakteId !== this.props.vkontakteId) nextState.vkontakteId = newProps.vkontakteId;
+      _this.setState({
+        internal: internal
+      }, internal.fetch);
+    });
 
-            if (newProps.skypeId !== this.props.skypeId) nextState.skypeId = newProps.skypeId;
+    _defineProperty(_assertThisInitialized(_assertThisInitialized(_this)), "_scaleTextNode", function (node) {
+      var _this$props = _this.props,
+          unstyled = _this$props.unstyled,
+          textSizeRatio = _this$props.textSizeRatio;
+      if (!node || unstyled) return;
+      var parent = node.parentNode; // Reset font-size such that scaling works correctly (#133)
 
-            if ((0, _keys2.default)(nextState) !== 0) {
-                nextState._internal = this.state._internal;
-                nextState._internal.sourcePointer = 0;
-                this.setState(nextState, this.fetch);
-            }
-        }
-    }, {
-        key: 'shouldComponentUpdate',
-        value: function shouldComponentUpdate(nextProps, nextState) {
-            return (0, _reactAddonsShallowCompare2.default)(this, nextProps, nextState);
-        }
-    }, {
-        key: '_renderAsImage',
-        value: function _renderAsImage() {
-            var size = this.props.size;
-            var round = this.props.round;
-            var alt = this.props.name || this.props.value;
-            var imageStyle = this.props.unstyled ? null : {
-                maxWidth: '100%',
-                width: size,
-                height: size,
-                borderRadius: round ? 500 : 0
-            };
-            return _react2.default.createElement('img', { className: this.props.innerClassName,
-                width: this.props.size,
-                height: this.props.size,
-                style: imageStyle,
-                src: this.state.src,
-                srcSet: this.state.srcSet,
-                alt: alt,
-                onError: this.fetch });
-        }
-    }, {
-        key: '_renderAsText',
-        value: function _renderAsText() {
-            var size = this.props.size;
-            var textSizeRatio = this.props.textSizeRatio;
-            var round = this.props.round;
-            var initialsStyle = this.props.unstyled ? null : {
-                width: size,
-                height: size,
-                font: Math.floor(size / textSizeRatio) + 'px Helvetica, Arial, sans-serif',
-                lineHeight: size + 'px', // yes, px suffix is needed on lineHeight
-                textAlign: 'center',
-                textTransform: 'uppercase',
-                color: this.props.fgColor,
-                background: this.state.color,
-                borderRadius: round ? '100%' : 0
-            };
-            return _react2.default.createElement(
-                'div',
-                { style: initialsStyle },
-                this.state.value
-            );
-        }
-    }, {
-        key: 'render',
-        value: function render() {
-            var size = this.props.size;
-            var hostStyle = this.props.unstyled ? null : (0, _extends3.default)({
-                display: 'inline-block',
-                width: size,
-                height: size,
-                borderRadius: this.props.round ? 500 : 0
-            }, this.props.style);
-            return _react2.default.createElement(
-                'div',
-                { className: this.props.className,
-                    style: hostStyle },
-                this.state.src ? this._renderAsImage() : this._renderAsText()
-            );
-        }
-    }]);
-    return Avatar;
-}(_react2.default.Component);
+      parent.style.fontSize = null;
+      var textWidth = node.getBoundingClientRect().width;
+      if (textWidth < 0) return;
+      var containerWidth = parent.getBoundingClientRect().width;
+      var ratio = containerWidth / textWidth; // Set font-size on parent span, otherwise the `table-cell` span
+      // will cause alignment issues.
 
-Avatar.displayName = 'Avatar';
-Avatar.propTypes = {
-    className: _propTypes2.default.string,
-    innerClassName: _propTypes2.default.string,
-    fgColor: _propTypes2.default.string,
-    color: _propTypes2.default.string,
-    colors: _propTypes2.default.array,
-    name: _propTypes2.default.string,
-    value: _propTypes2.default.string,
-    email: _propTypes2.default.string,
-    md5Email: _propTypes2.default.string,
-    src: _propTypes2.default.string,
-    srcSet: _propTypes2.default.string,
-    facebookId: _propTypes2.default.string,
-    googleId: _propTypes2.default.string,
-    twitterHandle: _propTypes2.default.string,
-    vkontakteId: _propTypes2.default.string,
-    skypeId: _propTypes2.default.string,
-    round: _propTypes2.default.bool,
-    style: _propTypes2.default.object,
-    size: _propTypes2.default.number,
-    textSizeRatio: _propTypes2.default.number,
-    unstyled: _propTypes2.default.bool
-};
-Avatar.defaultProps = {
-    className: 'sb-avatar',
-    innerClassName: '',
-    fgColor: '#FFF',
-    color: null,
-    name: null,
-    value: null,
-    email: null,
-    md5Email: null,
-    facebookId: null,
-    googleId: null,
-    twitterHandle: null,
-    vkontakteId: null,
-    skypeId: null,
-    round: false,
-    size: 100,
-    style: null,
-    textSizeRatio: 3,
-    unstyled: false
-};
-exports.default = Avatar;
-module.exports = exports['default'];
+      parent.style.fontSize = "calc((100% * ".concat(ratio, ") / ").concat(textSizeRatio, ")");
+    });
+
+    _this.state = {
+      src: null,
+      srcSet: null,
+      internal: null,
+      value: null,
+      color: props.color
+    };
+    return _this;
+  }
+
+  _createClass(Avatar, [{
+    key: "componentDidMount",
+    value: function componentDidMount() {
+      this.fetch();
+    }
+  }, {
+    key: "componentWillReceiveProps",
+    value: function componentWillReceiveProps(newProps) {
+      var needsUpdate = false; // This seems redundant
+      //
+      // Props that need to be in `state` are
+      // `value`, `src` and `color`
+
+      for (var prop in sourcePropTypes) {
+        needsUpdate = needsUpdate || newProps[prop] !== this.props[prop];
+      } // if ((newProps.srcSet !== this.props.srcSet) || (newProps.name !== this.props.name)) {
+      //     needsUpdate = true
+      // }
+
+
+      if (needsUpdate) setTimeout(this.fetch, 0);
+    }
+  }, {
+    key: "componentWillUnmount",
+    value: function componentWillUnmount() {
+      if (this.state.internal) {
+        this.state.internal.active = false;
+      }
+    }
+  }, {
+    key: "_renderAsImage",
+    value: function _renderAsImage() {
+      var _this$props2 = this.props,
+          className = _this$props2.className,
+          round = _this$props2.round,
+          unstyled = _this$props2.unstyled,
+          name = _this$props2.name,
+          value = _this$props2.value;
+      var internal = this.state.internal;
+      var size = (0, _utils.parseSize)(this.props.size);
+      var alt = name || value;
+      var imageStyle = unstyled ? null : {
+        maxWidth: '100%',
+        width: size.str,
+        height: size.str,
+        borderRadius: round === true ? '100%' : round
+      };
+      return _react.default.createElement("img", {
+        className: className + ' sb-avatar__image',
+        width: size.str,
+        height: size.str,
+        style: imageStyle,
+        src: this.state.src,
+        srcSet: this.state.srcSet,
+        alt: alt,
+        onError: internal && internal.fetch
+      });
+    }
+  }, {
+    key: "_renderAsText",
+    value: function _renderAsText() {
+      var _this$props3 = this.props,
+          className = _this$props3.className,
+          round = _this$props3.round,
+          unstyled = _this$props3.unstyled;
+      var size = (0, _utils.parseSize)(this.props.size);
+      var initialsStyle = unstyled ? null : {
+        width: size.str,
+        height: size.str,
+        lineHeight: 'initial',
+        textAlign: 'center',
+        textTransform: 'uppercase',
+        color: this.props.fgColor,
+        background: this.state.color,
+        borderRadius: round === true ? '100%' : round
+      };
+      var tableStyle = unstyled ? null : {
+        display: 'table',
+        width: '100%',
+        height: '100%'
+      };
+      var spanStyle = unstyled ? null : {
+        display: 'table-cell',
+        verticalAlign: 'middle'
+      };
+      return _react.default.createElement("div", {
+        className: className + ' sb-avatar__text',
+        style: initialsStyle
+      }, _react.default.createElement("div", {
+        style: tableStyle
+      }, _react.default.createElement("span", {
+        style: spanStyle
+      }, _react.default.createElement("span", {
+        ref: this._scaleTextNode,
+        key: this.state.value
+      }, this.state.value))));
+    }
+  }, {
+    key: "render",
+    value: function render() {
+      var _this$props4 = this.props,
+          className = _this$props4.className,
+          unstyled = _this$props4.unstyled,
+          round = _this$props4.round,
+          style = _this$props4.style,
+          onClick = _this$props4.onClick;
+      var _this$state = this.state,
+          src = _this$state.src,
+          sourceName = _this$state.sourceName;
+      var size = (0, _utils.parseSize)(this.props.size);
+      var hostStyle = unstyled ? null : _objectSpread({
+        display: 'inline-block',
+        verticalAlign: 'middle',
+        width: size.str,
+        height: size.str,
+        borderRadius: round === true ? '100%' : round,
+        fontFamily: 'Helvetica, Arial, sans-serif'
+      }, style);
+      var classNames = [className, 'sb-avatar'];
+
+      if (sourceName) {
+        var source = sourceName.toLowerCase().replace(/[^a-z0-9-]+/g, '-') // only allow alphanumeric
+        .replace(/^-+|-+$/g, ''); // trim `-`
+
+        classNames.push('sb-avatar--' + source);
+      }
+
+      return _react.default.createElement("div", {
+        className: classNames.join(' '),
+        onClick: onClick,
+        style: hostStyle
+      }, src ? this._renderAsImage() : this._renderAsText());
+    }
+  }]);
+
+  return Avatar;
+}(_react.PureComponent);
+
+exports.Avatar = Avatar;
+
+_defineProperty(Avatar, "displayName", 'Avatar');
+
+_defineProperty(Avatar, "propTypes", _objectSpread({}, sourcePropTypes, {
+  className: _propTypes.default.string,
+  fgColor: _propTypes.default.string,
+  color: _propTypes.default.string,
+  name: _propTypes.default.string,
+  value: _propTypes.default.string,
+  email: _propTypes.default.string,
+  md5Email: _propTypes.default.string,
+  src: _propTypes.default.string,
+  srcSet: _propTypes.default.string,
+  facebookId: _propTypes.default.string,
+  googleId: _propTypes.default.string,
+  twitterHandle: _propTypes.default.string,
+  vkontakteId: _propTypes.default.string,
+  skypeId: _propTypes.default.string,
+  colors: _propTypes.default.arrayOf(_propTypes.default.string),
+  round: _propTypes.default.oneOfType([_propTypes.default.bool, _propTypes.default.string]),
+  style: _propTypes.default.object,
+  size: _propTypes.default.oneOfType([_propTypes.default.number, _propTypes.default.string]),
+  textSizeRatio: _propTypes.default.number,
+  unstyled: _propTypes.default.bool,
+  cache: _propTypes.default.object,
+  onClick: _propTypes.default.func
+}));
+
+_defineProperty(Avatar, "defaultProps", {
+  className: 'sb-avatar',
+  innerClassName: '',
+  fgColor: '#FFF',
+  round: false,
+  size: 100,
+  textSizeRatio: 3,
+  unstyled: false
+});
+
+_defineProperty(Avatar, "getRandomColor", _utils.getRandomColor);
+
+_defineProperty(Avatar, "ConfigProvider", _context.ConfigProvider);
+
+var _default = Object.assign((0, _context.withConfig)(Avatar), {
+  getRandomColor: _utils.getRandomColor,
+  ConfigProvider: _context.ConfigProvider
+});
+
+exports.default = _default;
