@@ -1,38 +1,47 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-    value: true
+  value: true
 });
+exports.default = void 0;
 
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
-
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+
 var FacebookSource = function FacebookSource(props) {
-    var _this = this;
+  var _this = this;
 
-    (0, _classCallCheck3.default)(this, FacebookSource);
-    this.props = null;
+  _classCallCheck(this, FacebookSource);
 
-    this.isCompatible = function () {
-        return !!_this.props.facebookId;
-    };
+  _defineProperty(this, "props", null);
 
-    this.get = function (setState) {
-        var _props = _this.props,
-            size = _props.size,
-            facebookId = _props.facebookId;
+  _defineProperty(this, "isCompatible", function () {
+    return !!_this.props.facebookId;
+  });
 
-        var url = 'https://graph.facebook.com/' + (facebookId + '/picture?width=' + size + '&height=' + size);
-        var retinaUrl = 'https://graph.facebook.com/' + (facebookId + '/picture?width=' + size + '&height=' + size * 2);
+  _defineProperty(this, "get", function (setState) {
+    var _this$props = _this.props,
+        size = _this$props.size,
+        facebookId = _this$props.facebookId;
+    var url = 'https://graph.facebook.com/' + "".concat(facebookId, "/picture?width=").concat(size, "&height=").concat(size);
+    var retinaUrl = 'https://graph.facebook.com/' + "".concat(facebookId, "/picture?width=").concat(size, "&height=").concat(size * 2);
+    setState({
+      sourceName: 'facebook',
+      src: url,
+      srcSet: retinaUrl
+    });
+  });
 
-        setState({ src: url, srcSet: retinaUrl });
-    };
-
-    this.props = props;
+  this.props = props;
 };
 
 exports.default = FacebookSource;
-module.exports = exports['default'];
+
+_defineProperty(FacebookSource, "propTypes", {
+  facebookId: _propTypes.default.string
+});
