@@ -1,56 +1,66 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-    value: true
+  value: true
 });
+exports.default = void 0;
 
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+require("core-js/modules/es6.regexp.replace");
 
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _utils = require('../utils');
+var _utils = require("../utils");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+
 var GoogleSource = function GoogleSource(props) {
-    var _this = this;
+  var _this = this;
 
-    (0, _classCallCheck3.default)(this, GoogleSource);
-    this.props = null;
+  _classCallCheck(this, GoogleSource);
 
-    this.isCompatible = function () {
-        return !!_this.props.googleId;
-    };
+  _defineProperty(this, "props", null);
 
-    this.get = function (setState) {
-        var _props = _this.props,
-            size = _props.size,
-            googleId = _props.googleId;
+  _defineProperty(this, "isCompatible", function () {
+    return !!_this.props.googleId;
+  });
 
-        var url = 'https://picasaweb.google.com/data/entry/api/user/' + googleId + '?alt=json';
+  _defineProperty(this, "get", function (setState) {
+    var _this$props = _this.props,
+        cache = _this$props.cache,
+        size = _this$props.size,
+        googleId = _this$props.googleId;
+    var url = "https://picasaweb.google.com/data/entry/api/user/".concat(googleId, "?alt=json");
 
-        if ((0, _utils.hasSourceFailedBefore)(url)) {
-            setState(null);
-            return;
-        }
+    if (cache.hasSourceFailedBefore(url)) {
+      setState(null);
+      return;
+    }
 
-        (0, _utils.fetch)(url, function (data) {
-            var src = data.entry.gphoto$thumbnail.$t;
-            var srcWithCorrectSize = src.replace('s64', 's' + size);
-            var srcWithCorrectRetinaSize = src.replace('s64', 's' + size * 2);
-            setState({
-                src: srcWithCorrectSize,
-                srcSet: srcWithCorrectRetinaSize
-            });
-        }, function () {
-            // on error
-            (0, _utils.cacheFailingSource)(url);
-            setState(null);
-        });
-    };
+    (0, _utils.fetch)(url, function (data) {
+      var src = data.entry.gphoto$thumbnail.$t;
+      var srcWithCorrectSize = src.replace('s64', 's' + size);
+      var srcWithCorrectRetinaSize = src.replace('s64', 's' + size * 2);
+      setState({
+        sourceName: 'google',
+        src: srcWithCorrectSize,
+        srcSet: srcWithCorrectRetinaSize
+      });
+    }, function () {
+      // on error
+      cache.sourceFailed(url);
+      setState(null);
+    });
+  });
 
-    this.props = props;
+  this.props = props;
 };
 
 exports.default = GoogleSource;
-module.exports = exports['default'];
+
+_defineProperty(GoogleSource, "propTypes", {
+  googleId: _propTypes.default.string
+});
