@@ -1,47 +1,50 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-    value: true
+  value: true
 });
+exports.default = void 0;
 
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
-
-var _md = require('md5');
-
-var _md2 = _interopRequireDefault(_md);
+var _md = _interopRequireDefault(require("md5"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var GravatarSource = function GravatarSource(props) {
-    (0, _classCallCheck3.default)(this, GravatarSource);
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-    _initialiseProps.call(this);
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
-    this.props = props;
-};
+var GravatarSource = function GravatarSource(_props) {
+  var _this = this;
 
-var _initialiseProps = function _initialiseProps() {
-    var _this = this;
+  _classCallCheck(this, GravatarSource);
 
-    this.props = null;
+  _defineProperty(this, "props", null);
 
-    this.isCompatible = function () {
-        return !!_this.props.email || !!_this.props.md5Email;
-    };
+  _defineProperty(this, "isCompatible", function () {
+    return !!_this.props.email || !!_this.props.md5Email;
+  });
 
-    this.get = function (setState) {
-        var props = _this.props;
+  _defineProperty(this, "get", function (setState) {
+    var props = _this.props;
+    var email = props.md5Email || (0, _md.default)(props.email);
+    var size = props.size;
+    var url = "https://secure.gravatar.com/avatar/".concat(email, "?s=").concat(size, "&d=404");
+    var setUrl = "https://secure.gravatar.com/avatar/".concat(email, "?s=").concat(size * 2, "&d=404");
+    setState({
+      sourceName: 'gravatar',
+      src: url,
+      srcSet: setUrl
+    });
+  });
 
-        var email = props.md5Email || (0, _md2.default)(props.email);
-        var size = props.size;
-        var url = 'https://secure.gravatar.com/avatar/' + email + '?s=' + size + '&d=404';
-        var setUrl = 'https://secure.gravatar.com/avatar/' + email + '?s=' + size * 2 + '&d=404';
-
-        setState({ src: url, srcSet: setUrl });
-    };
+  this.props = _props;
 };
 
 exports.default = GravatarSource;
-module.exports = exports['default'];
+
+_defineProperty(GravatarSource, "propTypes", {
+  email: _propTypes.default.string,
+  md5Email: _propTypes.default.string
+});
