@@ -1,34 +1,42 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-    value: true
+  value: true
 });
+exports.default = void 0;
 
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
-
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
+
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
+
 var SrcSource = function SrcSource(props) {
-    var _this = this;
+  var _this = this;
 
-    (0, _classCallCheck3.default)(this, SrcSource);
-    this.props = null;
+  _classCallCheck(this, SrcSource);
 
-    this.isCompatible = function () {
-        return !!_this.props.src;
-    };
+  _defineProperty(this, "props", null);
 
-    this.get = function (setState) {
-        setState({
-            src: _this.props.src,
-            srcSet: _this.props.srcSet
-        });
-    };
+  _defineProperty(this, "isCompatible", function () {
+    return !!_this.props.src;
+  });
 
-    this.props = props;
+  _defineProperty(this, "get", function (setState) {
+    setState({
+      sourceName: 'src',
+      src: _this.props.src,
+      srcSet: _this.props.srcSet
+    });
+  });
+
+  this.props = props;
 };
 
 exports.default = SrcSource;
-module.exports = exports['default'];
+
+_defineProperty(SrcSource, "propTypes", {
+  src: _propTypes.default.string
+});
