@@ -1,76 +1,95 @@
 'use strict';
 
 Object.defineProperty(exports, "__esModule", {
-    value: true
+  value: true
 });
+exports.default = void 0;
 
-var _classCallCheck2 = require('babel-runtime/helpers/classCallCheck');
+require("core-js/modules/es6.function.name");
 
-var _classCallCheck3 = _interopRequireDefault(_classCallCheck2);
+var _propTypes = _interopRequireDefault(require("prop-types"));
 
-var _createClass2 = require('babel-runtime/helpers/createClass');
-
-var _createClass3 = _interopRequireDefault(_createClass2);
-
-var _utils = require('../utils');
+var _utils = require("../utils");
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var ValueSource = function () {
-    function ValueSource(props) {
-        var _this = this;
+function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError("Cannot call a class as a function"); } }
 
-        (0, _classCallCheck3.default)(this, ValueSource);
-        this.props = null;
+function _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if ("value" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }
 
-        this.isCompatible = function () {
-            return !!_this.props.name || !!_this.props.value;
-        };
+function _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }
 
-        this.get = function (setState) {
-            var value = _this.getValue();
-            var state = value ? {
-                value: value,
-                color: _this.getColor(value)
-            } : null;
-            setState(state);
-        };
+function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
-        this.props = props;
+var ValueSource =
+/*#__PURE__*/
+function () {
+  function ValueSource(props) {
+    var _this = this;
+
+    _classCallCheck(this, ValueSource);
+
+    _defineProperty(this, "props", null);
+
+    _defineProperty(this, "isCompatible", function () {
+      return !!(_this.props.name || _this.props.value || _this.props.email);
+    });
+
+    _defineProperty(this, "get", function (setState) {
+      var value = _this.getValue();
+
+      if (!value) return setState(null);
+      setState({
+        sourceName: 'text',
+        value: value,
+        color: _this.getColor()
+      });
+    });
+
+    this.props = props;
+  }
+
+  _createClass(ValueSource, [{
+    key: "getInitials",
+    value: function getInitials() {
+      var _this$props = this.props,
+          name = _this$props.name,
+          initials = _this$props.initials;
+      if (typeof initials === 'string') return initials;
+      if (typeof initials === 'function') return initials(name, this.props);
+      return (0, _utils.defaultInitials)(name, this.props);
     }
+  }, {
+    key: "getValue",
+    value: function getValue() {
+      if (this.props.name) return this.getInitials();
+      if (this.props.value) return this.props.value;
+      return null;
+    }
+  }, {
+    key: "getColor",
+    value: function getColor() {
+      var _this$props2 = this.props,
+          color = _this$props2.color,
+          colors = _this$props2.colors,
+          name = _this$props2.name,
+          email = _this$props2.email,
+          value = _this$props2.value;
+      var colorValue = name || email || value;
+      return color || (0, _utils.getRandomColor)(colorValue, colors);
+    }
+  }]);
 
-    (0, _createClass3.default)(ValueSource, [{
-        key: 'getInitials',
-        value: function getInitials() {
-            var name = this.props.name;
-            var parts = name.split(' ');
-            var initials = '';
-            for (var i = 0; i < parts.length; i++) {
-                initials += parts[i].substr(0, 1).toUpperCase();
-            }
-            return initials;
-        }
-    }, {
-        key: 'getValue',
-        value: function getValue() {
-            if (this.props.name) return this.getInitials();
-
-            if (this.props.value) return this.props.value;
-
-            return null;
-        }
-    }, {
-        key: 'getColor',
-        value: function getColor(value) {
-            var _props = this.props,
-                color = _props.color,
-                colors = _props.colors;
-
-            return color || (0, _utils.getRandomColor)(value, colors);
-        }
-    }]);
-    return ValueSource;
+  return ValueSource;
 }();
 
 exports.default = ValueSource;
-module.exports = exports['default'];
+
+_defineProperty(ValueSource, "propTypes", {
+  color: _propTypes.default.string,
+  name: _propTypes.default.string,
+  value: _propTypes.default.string,
+  email: _propTypes.default.string,
+  maxInitials: _propTypes.default.number,
+  initials: _propTypes.default.oneOfType([_propTypes.default.string, _propTypes.default.func])
+});
