@@ -1,9 +1,14 @@
 'use strict';
 
-import {fetch,cacheFailingSource,hasSourceFailedBefore} from '../utils';
+import PropTypes from 'prop-types';
+import {fetch} from '../utils';
 
 export default
 class GoogleSource {
+
+    static propTypes = {
+        googleId: PropTypes.string
+    }
 
     props = null;
 
@@ -14,10 +19,10 @@ class GoogleSource {
     isCompatible = () => !!this.props.googleId
 
     get = (setState) => {
-        const { size, googleId } = this.props;
+        const { cache, size, googleId } = this.props;
         const url = `https://picasaweb.google.com/data/entry/api/user/${googleId}?alt=json`;
 
-        if(hasSourceFailedBefore(url)) {
+        if (cache.hasSourceFailedBefore(url)) {
             setState(null);
             return;
         }
@@ -27,12 +32,13 @@ class GoogleSource {
             const srcWithCorrectSize = src.replace('s64', 's' + size);
             const srcWithCorrectRetinaSize = src.replace('s64', 's' + size*2);
             setState({
+                sourceName: 'google',
                 src: srcWithCorrectSize,
                 srcSet: srcWithCorrectRetinaSize
             });
         }, () => {
             // on error
-            cacheFailingSource(url);
+            cache.sourceFailed(url);
             setState(null);
         });
     }
