@@ -1,9 +1,22 @@
 'use strict';
 
-import {getRandomColor} from '../utils';
+import PropTypes from 'prop-types';
+import {getRandomColor, defaultInitials} from '../utils';
 
 export default
 class ValueSource {
+
+    static propTypes = {
+        color: PropTypes.string,
+        name: PropTypes.string,
+        value: PropTypes.string,
+        email: PropTypes.string,
+        maxInitials: PropTypes.number,
+        initials: PropTypes.oneOfType([
+            PropTypes.string,
+            PropTypes.func
+        ])
+    }
 
     props = null
 
@@ -12,19 +25,19 @@ class ValueSource {
     }
 
     isCompatible = () => {
-        return !!this.props.name || !!this.props.value;
+        return !!(this.props.name || this.props.value || this.props.email);
     }
 
-    getInitials()
-    {
-        const name = this.props.name;
-        const parts = name.split(' ');
-        let initials = '';
-        for(let i = 0 ; i < parts.length ; i++)
-        {
-            initials += parts[i].substr(0, 1).toUpperCase();
-        }
-        return initials;
+    getInitials() {
+        const { name, initials } = this.props;
+
+        if (typeof initials === 'string')
+            return initials;
+
+        if (typeof initials === 'function')
+            return initials(name, this.props);
+
+        return defaultInitials(name, this.props);
     }
 
     getValue() {
@@ -37,17 +50,22 @@ class ValueSource {
         return null;
     }
 
-    getColor(value) {
-        const {color, colors} = this.props;
-        return color || getRandomColor(value, colors);
+    getColor() {
+        const {color, colors, name, email, value} = this.props;
+        const colorValue = name || email || value;
+        return color || getRandomColor(colorValue, colors);
     }
 
     get = (setState) => {
         const value = this.getValue();
-        const state = value ? {
+
+        if (!value)
+            return setState(null);
+
+        setState({
+            sourceName: 'text',
             value: value,
-            color: this.getColor(value)
-        } : null;
-        setState(state);
+            color: this.getColor()
+        });
     }
 }
